refactor: view

+++ app/src/main/java/com/baseproject/util/view/View.kt
@@ -11,11 +11,9 @@ import androidx.core.view.marginTop
 import androidx.core.view.updateLayoutParams
 import androidx.core.view.updateMargins
 
-val Int.dp: Int
-    get() = (this / Resources.getSystem().displayMetrics.density).toInt()
+val Float.dp: Float get() = this / Resources.getSystem().displayMetrics.density
 
-val Int.px: Int
-    get() = (this * Resources.getSystem().displayMetrics.density).toInt()
+val Float.px: Float get() = this * Resources.getSystem().displayMetrics.density
 
 fun View.gone() {
     this.visibility = View.GONE
